refactor: unite data types in one file
