--- src/ts/types/date.ts
@@ -1,6 +1,0 @@
-interface IDate {
-  day: number,
-  year: number,
-}
-
-export default IDate;
