--- src/ts/types/exercise-result.ts
@@ -1,6 +1,0 @@
-interface IExerciseResult {
-  exercise: string,
-  score: number,
-}
-
-export default IExerciseResult;
